Move VCRTestDataDir and VCRDataDirTester into vcr/tester.py
vcr/__init__.py was ~440 lines of mixed class implementations and
re-exports. Extract the two classes and get_test_cases into a dedicated
tester.py module; __init__.py becomes a thin re-export layer (~70 lines).
All public import paths remain unchanged.
